+++ glia/src/workflow/llm_workflow.py
@@ -4,21 +4,33 @@ from .base_workflow import BaseWorkflow
 from .workflow_name import WorkflowName
 from glia.src.model.model_name import ModelName
 
+from glia.src.service.llm_service import LLMService
+
+
 
 class LLMWorkflow(BaseWorkflow):
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
+        self.resource_manager = self.resource_manager
+        if self.service is None:
+            self.service = LLMService(
+                name="LLMService",
+                call_model_name=ModelName.OPENCHAT,
+                resource_manager=self.resource_manager,
+            )
+        else:
+            if isinstance(self.service, LLMService):
+                pass
 
     async def execute(self):
 
         print(
-            f"Executing workflow {self.name.value} with resources: {self.call_model_resources}, start ..."
+            f"Executing workflow {self.name.value} with resources: {self.service.call_model_resources}, start ..."
         )
         await asyncio.sleep(1)
-        self.process_result = self.call_model(self.prev_result)
-        print(self.process_result)
+        self.process_result = await self.service(self.prev_result)
         print(
-            f"Executing workflow {self.name.value} with resources: {self.call_model_resources}, end ..."
+            f"Executing workflow {self.name.value} with resources: {self.service.call_model_resources}, end ..."
         )
         return self.process_result