+++ glia/src/workflow/ocr_workflow.py
@@ -6,6 +6,7 @@ from .base_workflow import BaseWorkflow
 class OCRWorkflow(BaseWorkflow):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
+        
 
     async def execute(self):
 
@@ -13,7 +14,7 @@ class OCRWorkflow(BaseWorkflow):
             f"Executing component {self.name.value} with resources: {self.call_model_resources}, start ..."
         )
         await asyncio.sleep(1)
-        self.process_result = self.call_model(self.prev_result)
+        self.process_result = self.service(self.prev_result)
         print(self.process_result)
         print(
             f"Executing component {self.name.value} with resources: {self.call_model_resources}, end ..."